Publish AS::Notifications even on errors
Currently, Coach is fairly agnostic about errors being raised within one
of the middlewares. However, this leads to the unfortunate behaviour
that error logging is abruptly terminated if the middleware raises an
exception.

This means that any metadata from the request, that we had intentionally
captured, is lost, and we only have request logs from the outer
application to rely on.

This commit changes Coach::Handler to always report telemetry, even if
the request raises an exception. The status will default to 500, which
is a common design choice in the kinds of applications we build.