Use instrument for coach.handler.finish event
This change uses ActiveSupport::Notifications.instrument to time and
publish the event for coach.handler.finish. This makes Coach consistent
with how it handles handler and middleware events. Such a change should
not break any integrations, and can therefore be included in a patch.

The upshot of this is that integrating Coach with third-party
instrumentation tools then becomes a small bit easier. This is because
using #instrument over #publish will trigger start and end events under
the same key, if your subscriber is listening appropriately.