Capture the start time for the whole request
Coach is very useful for capturing the duration of the middlewares
within the request, but does not yield any information about the start
of the first request (even though this is captured by RequestBenchmark).

This commit stores, and exposes, the initial start time of the request,
which can be used by downstream subscribers for more complex
calculations.