Bump gc_ruboconfig to v2.3.4
This bumps our dependency on `gc_ruboconfig` from v2.3.2 to
v2.3.4.

It make sense for this dependency to live in `Gemfile` rather
than in the gemspec because you only need it if you're explicitly
working on Coach, and *not* if you're in the development
environment of a project which itself depends on Coach.

I've also removed any explicit dependency on `rubocop`, since we
can rely on `gc_ruboconfig` to declare the right version of it as
a dependency.