[RFC] Support passing blocks to middleware
For some middleware, it might make sense to accept a block rather than a
hash (or a hash where one of the keys is a block). For example:

class Throttle < Coach::Middleware
  def initialize(context, next_middleware, config = {}, &block)
    @throttle = Rack::Attack::Throttle.new(..., &block)
    super
  end
end

This commit modifies the internals of Coach to pass that block through
the chain to the middleware. No specs yet!