Record status:0 and exception metadata when exception occurs
Coach makes an assumption that an unhandled exception deep in the
middleware chain is going to render a HTTP 500 response. However, how
your application handles that exception will ultimately affect the
response. In Rails applications it is common to have a multitude of Rack
middleware which do some additional response handling, in particular
catching certain classes of exception and remapping them. Coach
currently just bubbles the exception back to Rails' router, which
handles the exception in it's own way.

So, if you were using `finish_handler.coach` events in logs to check
response states, you might be misled i3nto thinking that a 500 error was
issued by the application when in fact it had been remapped into
something else by a Rack middleware further up the stack.

Therefore, this assumption is faulty, and I believe instead we should
explicitly mark the status as `0` under these circumstances (`nil` is an
alternative but not an integer which might cause havoc with other code
handlers). We also attach the exception itself which can be inspected by
other layers.